[Refactor] : 필드 타입 변경

+++ src/main/java/live/smoothing/sensordata/dto/goal/GoalHistoryResponse.java
@@ -14,6 +14,6 @@ import java.time.LocalDateTime;
 public class GoalHistoryResponse {
 
     private LocalDateTime date;
-    private Long goalAmount;
-    private Long amount;
+    private Double goalAmount;
+    private Double amount;
 }
