+++ src/main/java/live/smoothing/sensordata/dto/goal/GoalResponse.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GoalResponse {
-    private Long goalAmount;
+    private Double goalAmount;
     private Integer unitPrice;
 }
